+++ include/proxsuite-nlp/prox-solver.hxx
@@ -1,5 +1,6 @@
 /// @file
 /// @copyright Copyright (C) 2022 LAAS-CNRS, INRIA
+/// @copyright Copyright (C) 2024-2025 INRIA
 /// @brief     Implementations for the prox solver.
 #pragma once
 
@@ -228,7 +229,22 @@ template <typename Scalar>
 void ProxNLPSolverTpl<Scalar>::computeProblemDerivatives(
     const ConstVectorRef &x, Workspace &workspace, boost::mpl::true_) const {
   this->computeProblemDerivatives(x, workspace, boost::mpl::false_());
-  problem_->computeHessians(x, workspace, hess_approx == HessianApprox::EXACT);
+
+  switch (hess_approx) {
+  case HessianApprox::BFGS:
+    bfgs_strategy_.update(x, workspace.objective_gradient,
+                          workspace.objective_hessian);
+    break;
+
+  case HessianApprox::IDENTITY:
+    workspace.objective_hessian.setIdentity();
+    break;
+
+  default: // handle both EXACT and GAUSS_NEWTON
+    problem_->computeHessians(x, workspace,
+                              hess_approx == HessianApprox::EXACT);
+    break;
+  }
 }
 
 template <typename Scalar>